Ilyav 042019

+++ src/main/java/edu/ilyav/api/cotrollers/ExperienceController.java
@@ -14,29 +14,45 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/rest/private/experience")
-public class ExperienceController {
+public class ExperienceController extends BaseController {
 
 	@Autowired
 	private ExperienceService experienceService;
 
 	@RequestMapping(value = "/add", method = RequestMethod.POST)
-	public Experience add(@RequestBody Experience experience) throws ResourceNotFoundException {
-		return experienceService.saveOrUpdate(experience);
+	public Experience add(HttpServletRequest request, @RequestBody Experience experience) throws ResourceNotFoundException {
+		if(isGuestMode(request))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else {
+			return experienceService.saveOrUpdate(experience);
+		}
 	}
 
 	@RequestMapping(value = "/update", method = RequestMethod.PUT)
-	public Experience update(@RequestBody Experience experience) throws ResourceNotFoundException {
-		return experienceService.saveOrUpdate(experience);
+	public Experience update(HttpServletRequest request, @RequestBody Experience experience) throws ResourceNotFoundException {
+		if(isGuestMode(request))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else {
+			return experienceService.saveOrUpdate(experience);
+		}
 	}
 
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.DELETE)
-	public WebResponse delete(@PathVariable Long id) throws Exception {
-		return experienceService.delete(id);
+	public WebResponse delete(HttpServletRequest request, @PathVariable Long id) throws Exception {
+		if(isGuestMode(request))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else {
+			return experienceService.delete(id);
+		}
 	}
 
 	@RequestMapping(value = "/all", method = RequestMethod.GET)
-	public List<Experience> getAll() {
-		return experienceService.findAll();
+	public List<Experience> getAll(HttpServletRequest request) throws ResourceNotFoundException {
+		if(isGuestMode(request))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else {
+			return experienceService.findAll();
+		}
 	}
